% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{APIChanged}
\alias{APIChanged}
\title{API changed}
\usage{
APIChanged(message, correlation_id = NULL, status_code = 500)
}
\arguments{
\item{message}{The message}

\item{correlation_id}{The correlation ID}

\item{status_code}{The status code, default 500}
}
\description{
Raised when the API is not as expected
}
