% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asynchronous_auth.R
\name{AsyncAuth}
\alias{AsyncAuth}
\title{Initialise an AsyncAuth instance}
\usage{
AsyncAuth(environment)
}
\arguments{
\item{environment}{Env -- The environment.}
}
\value{
An AsyncAuth instance.
}
\description{
Initialise an AsyncAuth instance
}
\examples{
auth <- AsyncAuth(Environment$DEV)
}
