% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{AuthNotValid}
\alias{AuthNotValid}
\title{Auth not valid}
\usage{
AuthNotValid(message, correlation_id = NULL, status_code = 401)
}
\arguments{
\item{message}{The message}

\item{correlation_id}{The correlation ID}

\item{status_code}{The status code, default 401}
}
\description{
Raised when the provided auth token was not valid
}
