% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_auth.R
\name{BaseAuth}
\alias{BaseAuth}
\title{Initialise a new instance of BaseAuth}
\usage{
BaseAuth(environment)
}
\arguments{
\item{environment}{Env -- the environment.}
}
\value{
An instance of BaseAuth.
}
\description{
Initialise a new instance of BaseAuth
}
