% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_auth.R
\name{.BaseAuth}
\alias{.BaseAuth}
\title{Base authentication}
\description{
Base authentication

Base authentication
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{env}}{Env -- The environment.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{env}}{Env -- The environment.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{.BaseAuth$new()}}
\item \href{#method-is_valid}{\code{.BaseAuth$is_valid()}}
\item \href{#method-.is_in_time}{\code{.BaseAuth$.is_in_time()}}
\item \href{#method-clone}{\code{.BaseAuth$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialise an instance of BaseAuth
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.BaseAuth$new(environment)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{environment}}{Env -- The environment.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new BaseAuth object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_valid"></a>}}
\if{latex}{\out{\hypertarget{method-is_valid}{}}}
\subsection{Method \code{is_valid()}}{
Check if token is valid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.BaseAuth$is_valid(token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{token}}{The access token.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if token is valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.is_in_time"></a>}}
\if{latex}{\out{\hypertarget{method-.is_in_time}{}}}
\subsection{Method \code{.is_in_time()}}{
Check if token has expired.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.BaseAuth$.is_in_time(claims)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{claims}}{The claims}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if token has not expired.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.BaseAuth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
