% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{.decode_jwt}
\alias{.decode_jwt}
\title{Extract the claims from the token}
\usage{
.decode_jwt(token)
}
\arguments{
\item{token}{}
}
\value{
The claims
}
\description{
Extract the claims from the token
}
\keyword{internal}
