% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{.Env}
\alias{.Env}
\title{Defines an environment}
\description{
Defines an environment

Defines an environment
}
\keyword{internal}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{The name of the environment.}

\item{\code{base_url}}{The base URL for the environment.}

\item{\code{client_id}}{The client ID for the environment.}

\item{\code{audience}}{The audience for the environment.}

\item{\code{scope}}{The scope for the environment.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{.Env$new()}}
\item \href{#method-clone}{\code{.Env$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialise a new instance of the Env class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.Env$new(name, base_url, audience, client_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the environment.}

\item{\code{base_url}}{The base URL.}

\item{\code{audience}}{The audience.}

\item{\code{client_id}}{The client ID.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Env object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.Env$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
