% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{.get_platform_url}
\alias{.get_platform_url}
\title{Get the on-platform URL.}
\usage{
.get_platform_url()
}
\value{
The URL for on-platform.
}
\description{
Get the on-platform URL.
}
\keyword{internal}
