% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{.in_rstudio}
\alias{.in_rstudio}
\title{Determine whether this is running in an RStudio session on Datalab.}
\usage{
.in_rstudio()
}
\value{
TRUE if in an RStudio session.
}
\description{
Determine whether this is running in an RStudio session on Datalab.
}
\keyword{internal}
