% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_client.R
\name{.ProvenanceCache}
\alias{.ProvenanceCache}
\title{Provenance cache}
\description{
Provenance cache

Provenance cache
}
\details{
Encapsulate caching logic of provenance correlation ids
Designed to be inherited by clients
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{.cached_prov_corr_id}}{The cached provenance correlation ID.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{.cached_prov_corr_id}}{The cached provenance correlation ID.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{.ProvenanceCache$new()}}
\item \href{#method-.cache_response}{\code{.ProvenanceCache$.cache_response()}}
\item \href{#method-get_provenance_session_id}{\code{.ProvenanceCache$get_provenance_session_id()}}
\item \href{#method-clone}{\code{.ProvenanceCache$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialise a new instance of ProvenanceCache
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.ProvenanceCache$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.cache_response"></a>}}
\if{latex}{\out{\hypertarget{method-.cache_response}{}}}
\subsection{Method \code{.cache_response()}}{
Cache response
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.ProvenanceCache$.cache_response(headers)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{headers}}{Request headers}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_provenance_session_id"></a>}}
\if{latex}{\out{\hypertarget{method-get_provenance_session_id}{}}}
\subsection{Method \code{get_provenance_session_id()}}{
Return the Provenance Correlation Id of this session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.ProvenanceCache$get_provenance_session_id(token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{token}}{str -- The auth token of the current session}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
str -- the current Provenance Correlation Id, or NULL if there is
none present on the token and none has been cached yet
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.ProvenanceCache$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
