% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persisted_token.R
\name{.SimplePersistedToken}
\alias{.SimplePersistedToken}
\title{Simple file-based token persistence}
\description{
Simple file-based token persistence

Simple file-based token persistence
}
\details{
OK for scripts, not for web applications. Uses SyncAuth flow.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_token}{\code{.SimplePersistedToken$get_token()}}
\item \href{#method-clone}{\code{.SimplePersistedToken$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_token"></a>}}
\if{latex}{\out{\hypertarget{method-get_token}{}}}
\subsection{Method \code{get_token()}}{
Get an access token.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.SimplePersistedToken$get_token(environment, file_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{environment}}{Env -- The environment.}

\item{\code{file_path}}{str -- The file path for storing the access token.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An access token if available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.SimplePersistedToken$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
