% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\docType{data}
\name{Environment}
\alias{Environment}
\title{Environment}
\format{
An object of class \code{list} of length 5.
}
\usage{
Environment
}
\description{
Settings for environments
}
\keyword{datasets}
