% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_chaining_api_url}
\alias{get_chaining_api_url}
\title{Get chaining API URL}
\usage{
get_chaining_api_url(environment, path = "")
}
\arguments{
\item{environment}{Env -- The selected environment.}

\item{path}{str -- the path.}
}
\value{
Chaining API URL.
}
\description{
Get chaining API URL
}
