% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_client_id}
\alias{get_client_id}
\title{Get client ID}
\usage{
get_client_id(environment)
}
\arguments{
\item{environment}{Env -- The selected environment.}
}
\value{
The client ID for the environment.
}
\description{
Get client ID
}
