% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_headers}
\alias{get_headers}
\title{Get headers}
\usage{
get_headers(environment, token = "")
}
\arguments{
\item{environment}{Env -- The selected environment.}

\item{token}{str -- The access token.}
}
\value{
The headers.
}
\description{
Get headers
}
