% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_mme_api_url}
\alias{get_mme_api_url}
\title{Get MME API URL}
\usage{
get_mme_api_url(environment, path = "")
}
\arguments{
\item{environment}{Env -- The selected environment.}

\item{path}{str -- the path.}
}
\value{
MME API URL.
}
\description{
Get MME API URL
}
