% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_public_api_url}
\alias{get_public_api_url}
\title{Get public API URL.}
\usage{
get_public_api_url(environment, path = "")
}
\arguments{
\item{environment}{Env -- The selected environment.}

\item{path}{str -- the path.}
}
\value{
Public API URL.
}
\description{
Get public API URL.
}
