% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_scope}
\alias{get_scope}
\title{Get the scope}
\usage{
get_scope(environment)
}
\arguments{
\item{environment}{Env -- The selected environment.}
}
\value{
The scope for the environment
}
\description{
Get the scope
}
