% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_scopes}
\alias{get_scopes}
\title{Get scopes}
\usage{
get_scopes(environment)
}
\arguments{
\item{environment}{Env -- The selected environment.}
}
\value{
Scope as list.
}
\description{
Get scopes
}
