% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_token_url}
\alias{get_token_url}
\title{Get the URL to retrieve an access token}
\usage{
get_token_url()
}
\value{
The URL to poll for an access token
}
\description{
Get the URL to retrieve an access token
}
