% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{InputNotValid}
\alias{InputNotValid}
\title{Input not valid}
\usage{
InputNotValid(message, correlation_id = NULL, status_code = 400)
}
\arguments{
\item{message}{The message}

\item{correlation_id}{The correlation ID}

\item{status_code}{The status code, default 400}
}
\description{
Raised when the input data was not valid for the API call
}
