% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{InternalError}
\alias{InternalError}
\title{Internal error}
\usage{
InternalError(message, correlation_id = NULL, status_code = 500)
}
\arguments{
\item{message}{The message}

\item{correlation_id}{The correlation ID}

\item{status_code}{The status code 500}
}
\description{
Raised when the API expected to process the request but failed to do so
}
