% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_platform.R
\name{PlatformAuth}
\alias{PlatformAuth}
\title{Initialise a new instance of PlatformAuth}
\usage{
PlatformAuth()
}
\value{
An instance of PlatformAuth
}
\description{
Initialise a new instance of PlatformAuth
}
\examples{

# On platform - RStudio
env <- Environment$ON_PLATFORM
auth <- PlatformAuth()
token <- auth$get_token()

# On platform - Shiny app
env <- Environment$ON_PLATFORM
auth <- PlatformAuth()
token <- auth$get_token(session) # provide the shiny session

}
