% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_client.R
\name{ProvenanceCache}
\alias{ProvenanceCache}
\title{Initialise an instance of ProvenanceCache}
\usage{
ProvenanceCache()
}
\value{
A new instance of ProvenanceCache
}
\description{
Initialise an instance of ProvenanceCache
}
