% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{ResourceNotFound}
\alias{ResourceNotFound}
\title{Resource not found}
\usage{
ResourceNotFound(message, correlation_id = NULL, status_code = 404)
}
\arguments{
\item{message}{The message}

\item{correlation_id}{The correlation ID}

\item{status_code}{The status code, default 404}
}
\description{
Raised when a requested resource is missing
}
