% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persisted_token.R
\name{SimplePersistedToken}
\alias{SimplePersistedToken}
\title{Initialise an instance of SimplePersistedToken}
\usage{
SimplePersistedToken()
}
\value{
A SimplePersistedToken instance.
}
\description{
Initialise an instance of SimplePersistedToken
}
