% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchronous_auth.R
\name{SyncAuth}
\alias{SyncAuth}
\title{Initialise a new SyncAuth instance}
\usage{
SyncAuth(environment)
}
\arguments{
\item{environment}{Env -- The environment.}
}
\value{
A SyncAuth instance.
}
\description{
Initialise a new SyncAuth instance
}
