% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{TemporaryError}
\alias{TemporaryError}
\title{Temporary error}
\usage{
TemporaryError(message, correlation_id = NULL, status_code = 502)
}
\arguments{
\item{message}{The message}

\item{correlation_id}{The correlation ID}

\item{status_code}{The status code, default 502}
}
\description{
Raised when the API had a temporary issue in processing the request
}
